Fix Metrics/ModuleLength
